508965 - Rounding up obligation calculation - tonnage type from double to integer within classes

+++ src/EPR.PRN.Backend.Obligation/Services/MaterialCalculationService.cs
@@ -4,12 +4,12 @@ namespace EPR.PRN.Backend.Obligation.Services
 {
     public class MaterialCalculationService : IMaterialCalculationService
     {
-        public int Calculate(double target, double tonnage)
+        public int Calculate(double target, int tonnage)
         {
             return (int)Math.Ceiling(target * tonnage);
         }
 
-        public (int remelt, int remainder) CalculateGlass(double target, double remeltTarget, double tonnage)
+        public (int remelt, int remainder) CalculateGlass(double target, double remeltTarget, int tonnage)
         {
             var initialTarget = (int)Math.Ceiling(target * tonnage);
             var remelt = (int)Math.Ceiling(remeltTarget * initialTarget);
